+++ lib/types/index.d.ts
@@ -8,20 +8,91 @@ declare module 'react-native-wifi-reborn' {
         timestamp: number;
     };
 
-    export type Errors = Partial<{
-        // The WIFI network is not currently in range.
-        notInRange: boolean;
-        // Could not add or update the network configuration.
-        addOrUpdateFailed: boolean;
-        // Disconnecting from the network failed. This is done as part of the connect flow
-        disconnectFailed: boolean;
-        // Could not connect to network
-        connectNetworkFailed: boolean;
-    }>;
+    export enum CONNECT_ERRORS {
+        /**
+         * Starting from iOS 11, NEHotspotConfigurationError is available
+         */
+        unavailableForOSVersion = 'unavailableForOSVersion',
+        /**
+         * If an unknown error is occurred (iOS)
+         */
+        invalid = 'invalid',
+        /**
+         * If the SSID is invalid
+         */
+        invalidSSID = 'invalidSSID',
+        /**
+         * If the SSID prefix is invalid
+         */
+        invalidSSIDPrefix = 'invalidSSIDPrefix',
+        /**
+         * If the passphrase is invalid
+         */
+        invalidPassphrase = 'invalidPassphrase',
+        /**
+         * If the user canceled the request to join the asked network
+         */
+        userDenied = 'userDenied',
+        /**
+         * Starting from iOS 13, location permission is denied (iOS)
+         */
+        locationPermissionDenied = 'locationPermissionDenied',
+        /**
+         * When an unknown error occurred
+         */
+        unableToConnect = 'unableToConnect',
+        /**
+         * Starting from iOS 13, location permission is restricted (iOS)
+         */
+        locationPermissionRestricted = 'locationPermissionRestricted',
+        /**
+         * Starting android 6, location permission needs to be granted for wifi scanning.
+         */
+        locationPermissionMissing = 'locationPermissionMissing',
+        /**
+         * Starting Android 6, location services needs to be on to scan for wifi networks.
+         */
+        locationServicesOff = 'locationServicesOff',
+        /**
+         * Starting Android 10, apps are no longer allowed to enable wifi.
+         * User has to manually do this.
+         */
+        couldNotEnableWifi = 'couldNotEnableWifi',
+        /**
+         * Starting Android 9, it's only allowed to scan 4 times per 2 minuts in a foreground app.
+         * https://developer.android.com/guide/topics/connectivity/wifi-scan
+         */
+        couldNotScan = 'couldNotScan',
+        /**
+         * If the SSID couldn't be detected
+         */
+        couldNotDetectSSID = 'couldNotDetectSSID',
+        /**
+         * If the wifi network is not in range, the security type is unknown and WifiUtils doesn't support
+         * connecting to the network.
+         */
+        didNotFindNetwork = 'didNotFindNetwork',
+        /**
+         * Authentication error occurred while trying to connect.
+         * The password could be incorrect or the user could have a saved network configuration with a
+         * different password!
+         */
+        authenticationErrorOccurred = 'authenticationErrorOccurred',
+        /**
+         * Firmware bugs on OnePlus prevent it from connecting on some firmware versions.
+         * More info: https://github.com/ThanosFisherman/WifiUtils/issues/63
+         */
+        android10ImmediatelyDroppedConnection = 'android10ImmediatelyDroppedConnection',
+        /**
+         * Could not connect in the timeout window.
+         */
+        timeoutOccurred = 'timeoutOccurred',
+    }
 
     /**
      * Connects to a WiFi network. Rejects with an error if it couldn't connect.
      *
+     * @param SSID Wifi name.
      * @param password `null` for open networks.
      * @param isWep Used on iOS. If `true`, the network is WEP Wi-Fi; otherwise it is a WPA or WPA2 personal Wi-Fi network.
      */
@@ -84,22 +155,13 @@ declare module 'react-native-wifi-reborn' {
 
     /**
      * Returns a list of nearby WiFI networks.
-     *
-     * @example
-     * const results = await WifiManager.loadWifiList();
-        results => {
-            let wifiArray =  JSON.parse(results);
-            wifiArray.map((value, index) =>
-                console.log(`Wifi ${index  +  1} - ${value.SSID}`)
-            );
-        },
      */
     export function loadWifiList(): Promise<Array<WifiEntry>>;
 
     /**
      * Similar to `loadWifiList` but it forcefully starts a new WiFi scan and only passes the results when the scan is done.
      */
-    export function reScanAndLoadWifiList(): Promise<Array<string>>;
+    export function reScanAndLoadWifiList(): Promise<Array<WifiEntry>>;
 
     /**
      * Method to check if wifi is enabled.
@@ -113,7 +175,20 @@ declare module 'react-native-wifi-reborn' {
      */
     export function connectionStatus(): Promise<boolean>;
 
-    export function disconnect(): void;
+    export const DISCONNECT_ERRORS = {
+        /**
+         * Could not get the WifiManager.
+         * https://developer.android.com/reference/android/net/wifi/WifiManager?hl=en
+         */
+        couldNotGetWifiManager = 'couldNotGetWifiManager',
+        /**
+         * Could not get the ConnectivityManager.
+         * https://developer.android.com/reference/android/net/ConnectivityManager?hl=en
+         */
+        couldNotGetConnectivityManager = 'couldNotGetConnectivityManager',
+    };
+
+    export function disconnect(): Promise<boolean>;
 
     /**
      * Returns the BSSID (basic service set identifier) of the currently connected WiFi network.
@@ -140,6 +215,8 @@ declare module 'react-native-wifi-reborn' {
          * Starting android 6, location permission needs to be granted for wifi scanning.
          */
         locationPermissionMissing = 'locationPermissionMissing',
+        couldNotGetWifiManager = 'couldNotGetWifiManager',
+        couldNotGetConnectivityManager = 'couldNotGetConnectivityManager',
     };
 
     /**
@@ -155,17 +232,23 @@ declare module 'react-native-wifi-reborn' {
     }
 
     /**
-     * Use this to execute api calls to a wifi network that does not have internet access.
-     *
-     * Useful for commissioning IoT devices.
-     *
-     * This will route all app network requests to the network (instead of the mobile connection).
+     * @deprecated Use forceWifiUsageWithOptions.
+     */
+    export function forceWifiUsage(useWifi: boolean): Promise<void>;
+
+    /**
+     * Use this to route all app network requests to the wifi network (instead of the mobile connection).
      * It is important to disable it again after using as even when the app disconnects from the wifi
      * network it will keep on routing everything to wifi.
      *
-     * @param useWifi boolean to force wifi off or on
+     * Useful for commissioning IoT devices. If the wifi access point has no internet you can indicate so with
+     * the option `noInternet`.
+     *
+     * @param useWifi Force wifi usage on or off.
+     * @param options `noInternet` To indicate the access point has no internet. Usefull as some
+     * phone vendor customizations will switch back to mobile when the wifi access point has no internet.
      */
-    export function forceWifiUsage(useWifi: boolean): Promise<void>;
+    export function forceWifiUsageWithOptions(useWifi: boolean, options: { noInternet: boolean });
 
     //#endregion
 }